# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.Joke:
    properties:
      id:
        type: string
      joke:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: jeanyamazian@outlook.com
    name: Jean Yamazian
    url: https://jean.yamazian.com
  description: A high-performance REST API service for serving random jokes with built-in
    caching, rate limiting, and health checks.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Jokes Provider API
  version: "1.0"
paths:
  /api/v1/jokes/random:
    get:
      consumes:
      - application/json
      description: Returns a random joke from the jokes database. Supports caching.
      produces:
      - application/json
      responses:
        "200":
          description: Random joke object with id and joke fields
          schema:
            $ref: '#/definitions/models.Joke'
        "500":
          description: Failed to retrieve joke
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a random joke
      tags:
      - jokes
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"